+++ Runtime/Tweens/ITweenCommand.cs
@@ -1,0 +1,8 @@
+﻿namespace Timespawn.EntityTween.Tweens
+{
+    internal interface ITweenParams
+    {
+        void SetTweenParams(TweenParams tweenParams);
+        TweenParams GetTweenParams();
+    }
+}