+++ Runtime/Tweens/ITweenId.cs
@@ -1,0 +1,8 @@
+﻿namespace Timespawn.EntityTween.Tweens
+{
+    internal interface ITweenId
+    {
+        void SetTweenId(int id);
+        int GetTweenId();
+    }
+}