+++ Runtime/Tweens/ITweenInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace Timespawn.EntityTween.Tweens
+{
+    internal interface ITweenInfo<T>
+    {
+        void SetTweenInfo(T start, T end);
+        T GetTweenStart();
+        T GetTweenEnd();
+    }
+}